?for replaced with smart for
As promised, muforth now only has a smart for - one that loops 0 times
when given a loop count of 0.

I have edited the files in mu/lib/ and mu/test/ to conform with this
change, but to keep the changes modular, I haven't yet touched any of
the target metacompilers. In other words, they are sure to break, until
further changes get pushed.

I don't know why I didn't do this a loooong time ago. ;-)

NOTE: The C code has changed and must be recompiled.