Fixed number parsing problems unearthed by for/next change
The number parsing code is a mess. It's ugly, twisted, and a bit
embarrassing. It has gone thru a lot changes, and received various small
fixes over the years.

This is Yet Another Very Small Fix.

The change from dumb to smart for/next looping caused an awkward piece
of code with overlapping control structures - a for/next loop and an
if/then (written even more confusingly as a while/then) to fail. This
code was terrible in the first place, so it's good that it failed, like
a canary.

It is now infinitesimally better, and I've gone thru and re-written some
stack comments to be more consistent, and cleaned up one or two other
routines. The whole, however, is still a bit of a mess. At least it is a
bit more approachable, and it's possible to see a way towards a more
elegant solution.

Who knew parsing numbers was so hard? ;-)