Target meta-compilers updated to use smart for/next
I've changed the *host-side* code, which used to use ?for/next/then, to
use the new smarter for/next instead.

Note that this doesn't apply to the Forths that are being meta-compiled!
With the exception of the MSP430 - the impetus for this work - these
still need to be updated. I will do this as I touch the different
targets.

There is no reason why I lived so long with such a clumsy tool. Having a
smart for/next is much nicer.