msp430: A handful of self-contained files load the build env
Even though their names don't start with "load-" (which is usually how I
name files that are intended to first load the build environment, and
then load a project), there a few self-contained things that one might
want to run, and because all of these depend only on the basic build
environment and *not* on any of the kernel code, as a convenience (and
clarification) I decided they should load their own environment
automagically.

So to load any of these, just do this:

  ./muforth -f target/MSP430/<file>

and it should load all the other stuff too.

I hope this makes things simpler and better and does not add to
confusion - even though it kind of breaks a naming convention (that of
prefixing "load-" to the filename).