msp430: Made register printing smarter; Loop shows "dead" unless valid
Now when using "bug" inside of a loop, you'll see a reasonable value for
the Loop "register" (which may, in fact, be implemented by a stack slot
and not a machine register).

When not in a loop (really when executing some code with a "bug" in it),
Loop simply shows as "dead".