msp430: Made kernel a bit smaller by sharing wpush code
I noticed that a lot of kernel words - generally most words that push one or
more values onto the stack during their execution - ended in a similar
way:

  <get some value into a register>
  top push  <reg> top mov  next

That last line is a sequence of 4 words. If we instead compile a jump to
it, we can save 3 words (though in practice sometimes it works out to
only two words saved).

By doing this - by sharing the code that pushes w as a new value (in the
"create" runtime at the start of the kernel - look for "label wpush") in
several places I was able to save a whopping 34 bytes of code! Which
might not sound like much, but the kernel was only 544 bytes before
doing this, so it saved 6.25%!

Of course, there is a slight execution time penalty for this: the jmp
adds two cycles. Maybe it should be optional... But I think there is
still more space to be saved (once, eg, multi-tasking and the Forth chat
code is finished), so I'm going to leave it alone for now.

NOTE: This is a "breaking" kernel change. Re-flash the kernel before
proceeding.