Added a hook to file loading code to accumulate count of lines read
It's clunkier than doing it in C but nice to know that's is possible.

Basically, before closing a file we add the line number of the last
line, minus one (they start at 1 not 0) to a variable, lines-read.

After loading a bunch of files, just do

   lines-read @

to see how many.

Another hack was necessary. Since startup.mu4 is loaded in a slightly
different way, by simpler "bootstrap" C code, we have to add its length
by hand. We do this by adding a line at the end of startup.mu4 to do
just this!