msp430: asm: Fixed addressing mode operators - again!
Struggling to get this exactly right. The issue: In general it's nice to
convert index-by-0 to the shorter, faster indirect, if possible. The
trouble is that two special modes - indexing using sr as the base
register, and indexing using pc as the base - have special meanings. The
first is how absolute addresses are encoded; the second is how
pc-relative addresses are encoded. In both cases, converting to indirect
would change the meaning of the instruction.

My previous solution to this was simply to remove the optimization of
converting index-by-0 to indirect.

But sometimes you want it.

So now,  +)  (the indexing operator) is careful. If the index value is
zero AND the register is r4 to r15, convert to indirect. Otherwise,
leave it alone, and encode as indexed.

This seems to be both correct and efficient.