msp430: Added lsr, lsl, and rol synthetic instructions to asm
lsr is implemented using ror (a true 1op instruction) and can work with
any effective address.

lsl and rol are implemented using add and addc instructions, which are
2ops. The issue here is that the same effective address has to used for
source and dest. This gets sticky with some of the addressing modes,
like post-increment.

At the moment I've limited these to using only the register direct
addressing mode; this isn't a *necessary* limitation, but it's a helpful
simplification.