Linux console palette code can export nimblemachines.com palette
This is not something that most people are going to care about, but I
thought it might be useful for *me* - to be able to output the Linux
console palette colors in a form that I could easily post on my website.

Which I haven't done. But this code has been kicking around and needs a
good home. ;-)