Slightly improved how file loading works
Before, the stack depth check was only done for files loaded with
load-file, and skipped when using raw-load-file. This is wrong. The
stack depth should be checked in both cases, so I moved those checks
into raw-load-file.

Now the only difference between load-file and raw-load-file is that the
former saves some state before calling the latter.

I also tried to clean up the comments so they are more illuminating.