msp430: Small changes to fix the Hitachi HD44780 code
After having changed how Forth's NEXT works, I neglected to update a couple
of ;code words in this file.

I also realized that I was calculating loop delays wrong. Because I'm
using an immediate value of 1 in my delay loop - which can be generated
by the constant generator, so no actual immediate word is necessary - a
loop that decrements a register and jumps if not zero takes 3 cycles,
not 4. So I updated my loop count calculation for the delay words and
added a nice (but complicated) comment about it.

To update this code (assuming an up to date chat and the pre-HC595
version):

  ./muforth -f target/MSP430/load-lcd.mu4
  chat
  flash-image
  verify

If, as is likely, you've wired up the HC595 version, do this instead:

  ./muforth -d hc595 -f target/MSP430/load-lcd.mu4
  chat
  flash-image
  verify

Enjoy. I haven't tested the code myself. ;-)