msp430: Display of SP modified to show true top of stack
The displayed SP used to point to the pushed value of the Loop counter,
above which was the captured IP, and above *that* was the actual target
data stack.

Now the displayed SP "skips" the Loop and IP values on the stack and
instead points to the topmost data value. This should be less confusing
to users.