msp430: Changed to conditionally capture scratch VM registers
Minor changes to both the kernel trampoline/bug/chat code, and the
interact and register display code to conditionally capture the VM
scratch registers: W, X, Y, and Z.

This is intended to aid the process of getting familiar with how Forth
works on the MSP430, and is not intended for day-to-day use, though
there might be occasions when it could be helpful for debugging.

Because it was written as an optional "add-on" you have add

  -d prolix-bug

before loading the MSP430 build environment or the kernel, as it affects
both.