msp430: Upon reflection, we *can* show W - in most cases!
By modifying all the kernel's ;code words to capture W in a second
register - W' - which gets pushed in bug and shared with the host - we
can see into much more of the execution of the kernel. Because of the
way some kernel code shares code with the "create" runtime (search for
"label wpush" to see the definition, and "wpush" to see the uses), any
code using wpush will have its W copied into W' as well.

In order to use this special copying code - and the extra register
display - you have to define "microscope". I've done this (by default,
for now) in target/MSP430/load-kernel.mu4, but if you're loading the
kernel in another way, you need to have a

  -d microscope

somewhere *before* loading the build environment or the kernel code.

I hope this helps peer into the murk of a running Forth VM!