msp430: multdiv: take three of m*
Progress report on work-in-progress multiply and divide code
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
I made an assumption - which turned out to be false - in the code of
m*2, the second try of signed multiply. Unfortunately, while elegant,
the code is actually wrong in several cases involving the most negative
number - 8000 hex, in 16-bit land.

My assumption was that negating a negative number would always yield a
positive one. But because 2's complement has a singularity w.r.t. the
most-negative number - which negates to itself - this doesn't work.

It is safer, slower, and more correct to do what we did in the original
m* - check if the two multplicands have the same sign *first*, and start
making decisions based on that.

The only change really between m* and m*3 is that the latter takes
advantage of the fact that it knows that the multiplier is *negative*,
which means two things:

(1) the last step - the subtract - will always happen; no need to test
the bit; and

(2) because the result has to be negative (multiplicand is non-negative
and multiplier is negative), we can assume that the carry-in from the
left is always a one.

These two "optimizations" dull the pain - slightly - of the longer and
slower implementation compared to m*2.

NOTE: Just like m*, m*3 pushes a flag telling which way it decided to go
- mixed sign or same sign. This will go away when this code is ready for
prime time.