msp430: Disassembler defaults to postfix style! Can also do prefix.
After looking at the code and doing some deep thinking, I realised that
it wouldn't be that hard to make it flexible about the style of output
it creates.

It used to be that, unlike the assembler that you have to write - which
is postfix (Forth-style, with operands *preceding* the operators) - the
disassembler output text looked like more "traditional" assembler. This
is maybe nice in some ways - easier to share the output with someone who
isn't Forth-savvy - but it also means that you have to wear two hats
when doing assembler programming, constantly switching between these two
styles.

No longer!

If you prefer the postfix style for output, do nothing. It is now the
*default*.

If you prefer things the way they were (change is BAAAAAD!), then put
something like this:

  -d prefix-disasm

in target/MSP430/build.mu4, or maybe target/MSP430/load-kernel.mu4, or
some other file that you load to set up *your* preferred environment.