msp430: Added code to easily generate an Intex hex file from an image
Easiest way to use it (to generate a file) is like this:

  ./muforth -f target/MSP430/<loadfile> gen-ihex bye > <outputfile>

where <loadfile> is something like load-chat, load-kernel, or your
custom load file that loads the build environment, maybe the kernel, and
then some code.

This code has the same limitation that flash-image has: it can only
create an image of *one* flash region. So it's not great for general
purpose use, but to get a copy a copy of the chat code - for example, to
flash it onto a chip using MSPFlasher - do this:

  ./muforth -f target/MSP430/load-chat.mu4 gen-ihex bye > mspchat.hex