msp430: Added a "slow chat" flag to chat at 9600 bps
This was necessary to successfully talk to the (slow!) USB-to-serial
interface on the MSP-EXP430G2 LaunchPad. NOTE: I do *not* suggest
getting one of these! They are broken as designed. However! For anyone
stuck using one, it's possible to use it as a USB-to-serial bridge
between your PC (running muforth) and the chat code running on a G2xx
chip on the board.

Sadly, their USB-serial interface is limited to 9600 bps only, so I've
added a flag to the chat code to create a 9600 bps version. Load it like
this:

  ./muforth -d slow-chat -f target/MSP430/load-chat.mu4