msp430: Simplified memory region code
Now a region is a create/does> word whose runtime sets h - the
dictionary pointer pointer - to point to its body, which contain two
pointers: on origin, and a current pointer. (This part is not new. What
*is* new is that these underlying 2variable things were named; now they
are invisible.)

The only way to get access to the underlying pointers is by first
setting that region as the current one, by mentioning its name. Thus

  flash

will set the flash region as the current one.

  region

will push its start address and length. If you need to do something to a
particular region but don't want to change the notion of the current
region for everyone do this in your word:

  h preserve  <region>

where <region> is one of the defined regions: ram, flash, or info

h preserve saves the value of h and restores it when the words exits.