RISC-V support starts here! Initial commit of RISC-V skeleton...
Ok, this is very exciting day! Not only did House Republicans decide to
table their terrible healthcare legislation, but the beginnings of
RISC-V support have come to muforth!

There's not much you can do at the moment, but this commit will allow
you, assuming you have a version of openocd with RISC-V support, to
run openocd, attach to it, and read memory and registers.

You can probably change memory and write registers, but I haven't tested
that yet. Next comes the fun part: writing the assembler and
disassembler! But with this support in place I'll be able to test out
code as I go, on the chip!

How do you get started? It's pretty easy. You have to build or download
a copy of openocd that has RISC-V support. (How you do that is outside
the scope of this commit message. ;-)

Then edit the file

  mu/target/RISC-V/start-openocd.sh

so that the openocd variable points to *your* RISC-V-enabled openocd.

Then, in one windown, run that shell script.

In another window, try this:

  ./muforth -f target/RISC-V/build.mu4

then type

  gdb

This should connect to the gdb stub in openocd, through two named pipes
that get created in mu/target/RISC-V/.

Try

  @app du

Even more interesting is this address:

  @app 4e0 + du

Enjoy! And hopefully this will inflame your appetite for more goodies to
come!