riscv: Addition of a mostly-working bare-bones disassembler!!
It works (mostly) but is missing a few features:

* it disassembles rv32 I and M subsets, but not A or C. Unfortunately, a
lot of the code generated by the RISC-V gcc port uses C (compressed)
instructions. This is a *good* thing, but not for my disassembler. For
now it's rather hard to read code that includes subset C instructions.
This will change, but the compressed instructions are kinda complicated!
And I may tackle a few other things first...

* it doesn't name CSR regs, instead showing just their address. This
will change.

* it doesn't show "synthetic" (or "pseudo") instructions. Everything is
printed in the "raw" low-level form. This can be hard to read. But it is
also going to change. ;-)

* it doesn't (yet) calculate the addresses of jumps (for using 'g' and
'c' in the disassembler). This is an easy fix, and is also coming soon.
I would do it right now, but I've been working on this all day and want
to get something checked in that works!

Those are the main things missing.

One important note (esp if you are trying to compare the output to that
of gcc): all values printed are in hex.

Another important note: register name printing defaults to showing the
ABI names, but this can be switched at any time. raw-regs will show raw
names (x0 to x31); abi-names will show zero, ra, sp, etc.

One last thing: usage. As with all the dumpers/disassemblers included in
muforth, this one can be started in one of two ways:

  <addr> du

to start dumping words and bytes, or

  <addr> dis

to start disassembling. In the middle of disassembly 'd' will switch to
dumping, and from dumping 'i' gets you to the disassembler.

Enjoy!