riscv: disasm: Filled in the decodings; missing ops are defer'd words
The entire encoding space - with the exception of the larger width
(>32bit) instructions - is now decoded by the disassembler. The pieces
that are "missing" are all defer'd words that can be patched by later
code.

I'm still feeling the lack of C (compressed) disassembly as a big loss,
but it's a complicated beast with lots of crazy immediate/offset
encodings, and I want to tackle some other things before tackling this!