riscv: Added a Lua file that generates muforth defs for CSRs
The starting point is text "scraped" from the riscv-privileged spec PDF
file (v1.9.1).

The Lua code can be tweaked to generate whatever variations of the code
are necessary.

It "cleverly" (or stupidly, depending on your perspective) generates a
series of CSR definitions when it sees a definition in a particular
format. In the PDF file these series were represented by an ellipsis;
this code "unrolls" that ellipsis into a big pile of definitions.