core: .forth. and .compiler. are now true "muchain" words
Because they are defined in C and not Forth, .forth. and .compiler. were
previously a bit different from the dictionary chains created by the
word "chain", defined in startup.mu4.

Until now these differences were benign, mostly because the whole
purpose of doing this muchain stuff is to enable vocabulary chains to be
chained together, something I haven't been doing.

I've decided to dive into using chained vocabs, and I needed the
"native" ones to behave just like the Forth ones. This commit makes that
happen.

One observable behaviour that definitely changes with this commit is
that if you create a chain that is chained onto .forth., like this:

  .forth. chain .foo.

and then define a few words in .foo. :

  .foo. definitions
  : frog ;
  : turtle ;

Words will now show only the words actually in .foo., while *searching*
.foo. will find anything in .foo. or .forth. . I think this makes sense,
though it's maybe a bit surprising.

Try it out and see what you think.