core: Added a version of words - all-words - that continues thru muchain
Ok, so that "subject line" is possibly less than helpful. ;-)

The "problem" with the current version of words is that it stops when it
sees a "muchain" - the marker that it's joining another vocab chain.
Sometimes it's nice to see all the words that find would see if a chain
was searched. all-words will do this.