core: lib/clashes.mu4 updated to work with new "words" code
lib/clashes.mu4 can be used to find name "clashes" (collisions) between
two vocab chains.

Its use is a bit subtle, because the two chains are enumerated
differently. The first is enumerated, using the words code, only up to
the point that it joins another chain (if it is linked to another
chain). The second is enumerated by find, which always enumerates thru
any joins all the way to the root.

If you want to find the clashes with the words that exist only *in* a
chain - and not in any chains it links to - then specify it *first*.

To see this asymmetry in action, try this:

  ./muforth -f target/MSP430/build.mu4 -f lib/clashes.mu4
  .assembler. .forth. clashes
  .forth. .assembler. clashes

and note the difference. Because .assembler. is linked to .meta. - which
is in turn linked to .lex. - in the second case you see the clashes
between the .assembler. + .meta. + .lex. and .forth., and not just the
clashes between .assembeler. and .forth. . So take care with this!