riscv: Beginnings of a meta compiler and Forth kernel!
After spending a lot of time noodling around, trying to combine aspects
of the MSP430 meta-compiler and kernel with the ARM/v5/ meta-compiler
and kernel, and thinking that I wanted to do a *native* Forth, I finally
- today - decided to try instead a totally traditional ITC
(indirect-threaded code) implementation instead.

I haven't completed the code that will allow me to execute any of this
code on the target, so it is totally untested still - though I did find
an important bug in the assembler in the process. ;-)

I also am currently compiling NEXT inline, even though it's a bit of a
memory hog this way: each NEXT takes four instructions, or 16 bytes
total. Even with the tiny beginnings of a kernel, it already takes over
1 KiB of space. ;-)

I could easily share a single copy by jumping to it, and I may make this
change.

But I'm pretty excited to have made this start - finally!!