riscv: Fixed a few bugs related to pfa vs cfa
So, a word of notice/warning: I'm doing something that might seem
unorthodox, and yet it's not unprecedented. (Evidence of a similar
"trend" in Forth's existed from the early days - eg, in fig-FORTH.)

What am I talking about? I'm talking about the fact that target words
are "represented" (in the host dictionary, and in colon word bodies)
*not* by their cfa (code field address) but by their pfa (parameter
field address).

To give a concrete example, the word 0 is defined as a constant. This
means it has two parts: a code field, which points to the "doconst"
runtime; and a body, which consists of the 32-bit value 0000_0000. When
executed, it jumps to doconst, which pulls out its value and pushes it
onto the data stack.

If the cfa (code field address) of 0 is 8000_0008 (because we are
compiling into the RAM of the FE310), the pfa - which contains the value
- is the next cell, at address 8000_000c. It has been a tradition in
many Forths to use the cfa to represent a word by its cfa. ' (tick) will
often return a cfa, and colon word bodies will contain cfa's.

This sometimes makes NEXT a bit cleaner and faster, but sometimes it
makes doconst, dovar, docolon and friends uglier and slower.

I have decided in this case - since NEXT doesn't change in size or speed
- to use pfa's as the return value of ' and in colon word bodies. This
makes using ' to peer into words rather easier (you no longer have to
add an offset to skip the code field).

Because this unlike other Forths that I've meta-compiled using muforth,
there were some subtle bugs . I fixed at least one in this commit. ;-)