riscv: kernel: Shared NEXT saves about 256 bytes out of 1 KiB
So it might be worth doing. The key parts of the runtime - doconst,
docolon etc - have their copies of NEXT for speed, but the rest of the
code words that make up the kernel share a single copy and simply jump
to it instead of including it inline. For every jump that replaces NEXT,
12 bytes are saved.

It's also very easy to turn this on and off.