riscv: Made "safe" chatting mode easier to use
When your prompt says (chatting) you are connected to a target, ready to
remotely execute code.

In this mode also having access to the host's .forth. vocab chain can
cause problems and surprises. OTOH, there are a few words in .forth.
that are rather useful.

I've shut off access to .forth. for target/RISC-V/. To mitigate the
annoyance, I've added a handful of words to .meta. - to change the
radix, to explore the vocab chains, and to print numbers and reset the
stack. I hope this helps.