core: Moved old commented out Forth code to boneyard
For years now, dict.c and interpret.c have ended with a chunk of
commented-out Forth code. I think the idea was that this code roughly
represented what the C code was doing.

Since this code no longer accurately represents what the C code does, I
decided to move it out of the way.

I didn't want to toss it entirely, so I moved it into mu/lib/boneyard.mu4.