riscv: kernel: Added d+ d- dnegate d< ud< and host-side support
These words implement double-number addition, subtraction, and
comparison (signed and unsigned).

Since double numbers take two stack slots on the target (which for now
is 32-bit) and since the host machine has 64-bit stack slots, let's make
it easy to move double numbers back and forth.

This commit adds the following three meta-compiler words:

  >d  ( d - lo hi)  ( "to-d")
      Split a 64-bit host value into two 32-bit values. These latter are
      what the target code will operate on.

  d.  ( lo hi)
      Print a target double number - still split into two pieces - as a
      signed number.

  ud.  ( lo hi)
      Print a target double number - still split into two pieces - as an
      unsigned number.

An example of use would be something like this:

  -1 >d  -2 >d  d+  d.

Or, to compare two double-number values:

  -1 >d  4 >d  d<   ( signed less than)
  -1 >d  4 >d  ud<  ( unsigned less than)

NOTE: For this commit I have ifdef'ed out several target words - the
small constants 0 1 2 -1 etc in particular. These interfered with the
double-number code. I am currently pondering a change to both the host
muforth and to the target to corral certain words into a "runtime"
chain. These words will *not* be found and executed when interpreting;
they can only be *compiled*. This will help people not make certain
classes of mistake that could cause the system to blow up in their face,
like executing  >r  on the command line.