core: Moved a handful of words into a .runtime. chain
The reasoning is simple: These are words that should only be *compiled*,
never run in interpret mode.

By creating a new chain and moving these "compile-only" words into it,
we not only clean up the .forth. chain, but also make it impossible to
do simple foot-shootings like executing  >r  while interpreting.

This required some changes to the C names of routines and to the
dictionary code, so a complete re-configure and re-make is necessary.