riscv: Target "runtime" words are now in a separate chain
Like on the host, compile-only words are now placed onto
.target-runtime. and won't get accidentally executed while chatting.

On this list or also the small constants (0, 1, -1 etc), small constant
incrementers (like 1+), and the basic stack manipulation words.

The latter were added to make debugging easier. If you need to swap or
dup something on the stack, and you've executed "bug" and are trying to
look around, typing "dup" and having the _target_ dup get executed
(which will blow away your debug context) is annoying at best. So, it's
less likely now.

The small constants interfere with the use of double numbers on the
command line, and their whole reason for being is to make compiled code
shorter. They add nothing to the "interactive" experience.