riscv and msp430: Unchained .assembler. from .meta.
While in general chaining vocab chains together works well, with the
assembler we have to be careful. Because there are structure words -
if/then/else, begin/until etc - in both the .assembler. and .compiler.
chains, and because we definitely want to find the assembler's versions
*first*, we have to put .assembler. before .definer./.compiler. when
compiling inline assembler..

But there is a subtle problem. What if we decide, in some inline
assembler code, that we want to specify the use of a .forth. word, say,
swap. If we use \f - hoping to get the version from .compiler. - and if
.assembler. is chained to .meta. - we will fail. While searching
.assembler. for \f we will chain into .meta. and find its \f, which will
compile itself rather than the following word.

The easiest solution is simply to unchain .assembler. from .meta., and
search .meta. explicitly. We have to do that both in __inline-asm and in
__asm.)

I found this bug in the MSP430 code *after* changing it to use chained
vocabs. I decided that porting the fix to RISC-V to prevent inadvertent
failures there seemed wise.