msp430 riscv: Changed how target and target-runtime work
In a previous commit (5b3e8e423e48) I changed the meta-compiler to
create new target names in the current vocab chain, instead of in
.target. This was to facilitate putting *some* target words - those that
were deemed "compile-only" - onto a separate chain: .target-runtime.

The problem with this approach is that if the user happens to be in
.forth. (ie, current points to .forth. rather than to .target.) - which
used to often be the case when chatting - the new target word would get
created in .forth. and confusion would ensue. (In testing another change
this happened to me, so I figured it was a problem.)

To solve this, and to return to a less confusion-creating state of
affairs, target names are once again *always* created in .target.,
regardless of the setting of current. To get a name into
.target-runtime. you *annotate* its definition with [r] . I've added
these annotations to the MSP430 and RISC-V kernels.