riscv: Changed allocation of stacks in RAM
There is a sort of Forth "standard" in how the return (R) and data (D)
stacks should be allocated. It largely stems from how multitasking has
traditionally been done in Forth. Each task has a "user area", R stack,
and D stack. It looks something like this (with higher memory addresses
at the top of the page):

          +-------------+
          |  user area  |
          |     SP      |
          |     SP0     |
  RP0 --->+-------------+<--- UP (user pointer)
          |             |
          |   R stack   |
          |             |
          +-------------+<--- RP (top of R stack)

             undefined

  SP0 --->+-------------+
          |   D stack   |
          |             |
          |     RP      |
          +-------------+<--- SP (top of D stack)

RP0 (sometimes called R0) and SP0 (sometimes called S0) point to the
*bottom* of the R stack and D stack, resp.

When a task is de-scheduled, the user area contains the task's SP0 (D
stack bottom) and SP (D stack top) pointers, and RP has been pushed onto
the D stack.

I only mention this because I was allocating the stacks in the other
order for RISC-V, and I've since changed this. Now the R-stack is at the
top of RAM (there is no user area yet, because we are single-tasking),
and the D stack is located below that.