riscv: Added double stack words, and timer/counter fun!
The double stack words got added because at first I was thinking I
needed them because I was going to be computing with the 64-bit values
coming out of the RISC-V performance counters.

It turns out that the interesting experiment I ended up trying only
needed the low 32 bits, so things got simpler, but I had written the
double stack words - 2dup 2swap 2over and 2tuck - and decided to leave
them in the kernel.

The interesting stuff is the code that reads the counters (in
particular, mcycle and minstret) several times in quick succession and
calculates the differences. Some interesting cache behaviour - which I
haven't totally analyzed yet - shows up.

There are four experiments to try:

   read 32 bits from counter      \ /     shared NEXT
   read 64 bits from counter      / \     inline NEXT

Some quick background: The target Forth runs threaded code, which means
that every machine code routine ends with a "dispatch" routine. Shared
NEXT means that there is *one* version of this code in the kernel, and
every machine code word ends with a *jump* to the shared NEXT. Inline
NEXT means every machine code words ends with the four instructions of
NEXT. Shared NEXT code will obviously save some space, at the expense of
running slightly slower.

The first time I tried a different version of this code I got surprising
results: the shared NEXT code was way faster than the inline NEXT code.
After thinking about this I decided it had to do with the instruction
cache.

Bit of background #2: Every time you execute a word on the target, a
fetch.i instruction is executed. I decided this was the only safe thing
to do. The target has no idea what memory might have changed since it
last executed any code.

So any code execution starts with a "cold" i-cache.

Bit of background #3: Since this is a 32-bit Forth, 64-bit values are
represented on the stack by two values: the higher on top, and the lower
below it.

To load the shared NEXT version of the code, do this:

  ./muforth -d shared-next -f target/RISC-V/build.mu4

To load the inline NEXT version, leave off the -d shared-next. Try the
following experiments both ways, for maximum pleasure.

So here are the words:

  4c     reads the mcycle counter 4 times and pushes the results
  4cc    reads all 64 bits of the cycle counter 4 times

  4i     reads the minstret counter 4 times
  4ii    reads all 64 bits of the minstret counter 4 times

We won't need the high words of the 64 bits reads to calculate our
differences, we need an easy way to throw them away.

  4lo    takes 4 64 bit values and drops the high words

Lastly, there is a word to calculate the 3 differences between 4 values:

  4diff

So, to read mcycle four times and see the three differences, do this:

  4c 4diff

To do this with the 64-bit values from the counter, do this:

  4cc 4lo 4diff

You'll see that the first difference is much bigger than the second two.
This is the i-cached showing up.

Try the same thing with 4i and 4ii.

Have fun!