riscv: Native version of RISC-V kernel! ("quasi-threaded"?)
This commit tweaks the meta-compiler and kernel to generate what could
be called "native" code, but should really be called something like
"quasi-threaded code" (QTC).

Yes, there is no threading. But the structure of the "native" kernel is
identical to the structure of the ITC kernel. The difference is that
code fields contain "jal" instructions, and NEXT is a single "jr"
instruction.

It's exactly the system that I suggested as a "thought experiment" in my
threaded code intro: http://muforth.nimblemachines.com/threaded-code/

There are leaf (code) routines that make no calls; and "twig" (colon)
routines that *only* make calls. "Call" in this context meaning "jal".

After I got the native version working, I merged it back into the ITC
version, trying to minimize the differences. I realized something
curious: during the execution of code words - especially the kernel
runtime words like (branch) and (lit) - the "ra" register behaves
*exactly* like the IP register in threaded code! In fact, for the QTC
version I made an alias for ra called ip so that I could re-use
unchanged the bulk of the threaded code kernel.

Because RISC-V is a branch-and-link architecture, a "native" RISC-V
kernel is more similar to an ITC kernel than would be the case for a
CISC call/return machine.

The nature of the jal instruction is to capture the address of the
following word in a register. In our runtime we can use this register to
fetch a literal value, a branch destination, or a variable's current
value, and if the value loaded was "inline" (as is the case with
literals and branches) we can skip it and return by using an offset from
the link register.

Also, because we can specify the link register for every jal, we can use
one link register for "threading" (ra), and another for code fields (w).

I haven't tried to measure the performance of this version versus the
ITC version. I would love to see the numbers. Doing things this way
involves a lot of jal's to jal's, incurring pipeline reloads over and
over. How bad is it?

Here's how to try it out. There are now *three* ways to load the RISC-V
kernel:

  ./muforth -d native -f target/RISC-V/build.mu4

will load the latest "native" or QTC version.

  ./muforth -f target/RISC-V/build.mu4

will load the inline NEXT ITC version.

  ./muforth -d shared-next -f target/RISC-V/build.mu4

will load the shared NEXT ITC version.

The size numbers are interesting. The QTC version is 1500 bytes; the
shared NEXT version is 1800; and the inline NEXT version is 2200.