riscv: Added a totally simple and bogus "benchmark"
I'm curious about execution overhead for various versions of threading:

* ITC inline NEXT
* ITC shared NEXT
* jal-style subroutine "threading" (I called this quasi-threaded
  elsewhere)

By executing a serious of nested loops that execute a small variety of
types of Forth words - code words, a variable, a create/does word, a
simple colon word - and by doing the inner loop 200,000 times, we can
get an idea of the overhead of the execution.

To use this code, try this:

  ./muforth -f target/RISC-V/build.mu4 -f target/RISC-V/bogus-bench.mu4

Fire up openocd in another window, then run

  jtag
  bench

bench leaves two values on the stack: the count of _cycles_ elapsed, and
the count of instructions retired. The ratio is interesting. Print it by
executing

  .ratio

which will print a*1000/b in decimal. Pretend the comma is a decimal
point.

You should see a value like 1,527. So our code is taking approx 1.5
cycle on average to execute each instruction.

That was the "inline NEXT" version. Let's try the shared NEXT version.
Load it like this:

  ./muforth -d shared-next -f target/RISC-V/build.mu4 -f target/RISC-V/bogus-bench.mu4

Run openocd as before. Then:

  jtag
  bench
  .ratio

I get 1,544.

Now for the really interesting one: "native" code. Or, threading using
jal instructions exclusively.

  ./muforth -d native -f target/RISC-V/build.mu4 -f target/RISC-V/bogus-bench.mu4

Run openocd as before. Then:

  jtag
  bench
  .ratio

Again, I get 2,146. A much worse ratio. All those jal's to jal's are
causing lots of pipeline refills. As expected. But here is what's
interesting. If we compare the _cycle_ count between inline NEXT and
native, something interesting emerges. If I copy (by hand) the cycle
count from the first run (I got 252_4c6a hex) and the native run (I got
25d_1d9f) and compare them with .ratio:

  25d_1d9f ( native)  252_4c6a ( inline NEXT)  .ratio

I get 1,018. So only 1.8% more cycles for the much simpler (and smaller)
native version.

The ratio of _instructions retired_ between those two is also
interesting:

  185_1d85 ( inline NEXT)  119_f6ec ( native)  .ratio

yields 1,380. The threaded code is a bit faster, but executes 38% _more_
instructions!